requirements.txt
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyMySQL==1.0.2
SQLAlchemy==1.4.29
Werkzeug==2.0.3
requests==2.26.0
pyjwt==2.4.0
python-dotenv==0.21.0
